Clean up of tool
Added newer logging lib
Using scanner to read input
Renamed root command
Removed exit in place of return
