+++ go.mod
@@ -4,8 +4,10 @@ go 1.12
 
 require (
 	github.com/ONSdigital/go-ns v0.0.0-20191104121206-f144c4ec2e58
+	github.com/ONSdigital/log.go v0.0.0-20191127134126-2a610b254f20
 	github.com/fatih/color v1.7.0
 	github.com/google/go-github/v28 v28.1.1
+	github.com/hokaccha/go-prettyjson v0.0.0-20190818114111-108c894c2c0e // indirect
 	github.com/johnnadratowski/golang-neo4j-bolt-driver v0.0.0-20181101021923-6b24c0085aae
 	github.com/mattn/go-colorable v0.1.4 // indirect
 	github.com/mattn/go-isatty v0.0.10 // indirect
